pyproject.toml
[project]
name = "fasthtml-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.9.1",
]
